Extract src/props.d.ts
