+++ src/props.d.ts
@@ -1,0 +1,16 @@
+import { Language } from "models.d";
+
+export interface ISnippetProps {
+  _id: number;
+  properties: {
+    title: string;
+  };
+}
+
+export interface IFragmentProps {
+  _id?: number;
+  properties: {
+    content?: string;
+    language?: Partial<Language>;
+  };
+}
